Generate report (#3)
* Genereate report geral - WIP

->Added report class - WIP
->Added template file
->Lacks test

* Generate Report Geral - HTML WORKING

->Created a report geral html file

* Relatorio fiscal - HTML WORKING

->Can generate Report fiscal in html file
->Added method loadUsuario in usuario class

* New Report Screen - WIP

-> Major buttons working
-> Adapter for endereço is working
-> Must implement remaining ones and return button

* Fechados - WIP

->Started implementation of Fechados fragments
->Area Fechada fragment is done
->Lote Fechado fragment is WIP

* Lote fechados - DONE

->Can load lotes and entradas fechadas
->New functions to handle reload of areas infos
->PDF convertion still not implemented
->Must test Lote deletion upon closing

* fixes

* Bug fixes

->Fixed bug in area and lote alter dialogs which crashed upon save without selected image
->Fixed bug in report for Mao de obra conversion
->Fixed C_Entrada_DeleteAllByLote which was not deleting entradas and entrada_detalhe rows

* Charts fixes - DONE

-> Changed charts colors
->Changed chart despesa to balanço with negative values

* added reload user info function

->New function for reloading user info

+++ app/src/main/java/com/trabalho/tg/View/Detalhe/AreaCriaAlterDialog.kt
@@ -28,6 +28,7 @@ import android.os.Environment
 import android.util.Log
 import androidx.core.content.FileProvider
 import com.trabalho.tg.Helper.cameraUtils
+import kotlinx.android.synthetic.main.fragment_lote_alter_dialog.*
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
@@ -260,7 +261,7 @@ class AreaCriaAlterDialog : Fragment() {
         fun saveBitmap(area : Area) {
             val cw = ContextWrapper(context)
             val directory = cw.getDir("imageDir", Context.MODE_PRIVATE)
-            val file = File(directory, area.ar_nome.replace(" ","_") + ".jpg")
+            val file = File(directory, "areaID_" + area.ar_id + "_" + area.ar_nome.replace(" ","_") + ".jpg")
             //if (!file.exists()) {
                 Log.d("path", file.toString())
                 image_path = file.toString()
@@ -307,22 +308,25 @@ class AreaCriaAlterDialog : Fragment() {
                     area.ar_nome = edtTxtNome_AreaDialog.text.toString()
                     area.ar_del = "A"
 
-                    bitmap = (imgArea_DialogFragment.drawable as BitmapDrawable).bitmap
+                    if(imgArea_DialogFragment.drawable.constantState != resources.getDrawable(R.drawable.ic_menu_camera).constantState) {
+                        bitmap = (imgArea_DialogFragment.drawable as BitmapDrawable).bitmap
 
-                    if(!area.ar_nome.equals(areaPam!!.ar_nome) && areaPam!!.ar_imagem != null){
-                        var file = File(areaPam!!.ar_imagem)
-                        if(file.exists()){
-                            try{
-                                file.delete()
-                            }
-                            catch(e : Exception){
-                                System.out.println("File not deleted" + e.message)
+                        if (!area.ar_nome.equals(areaPam!!.ar_nome) && areaPam!!.ar_imagem != null) {
+                            var file = File(areaPam!!.ar_imagem)
+                            if (file.exists()) {
+                                try {
+                                    file.delete()
+                                } catch (e: Exception) {
+                                    System.out.println("File not deleted" + e.message)
+                                }
                             }
                         }
+                        saveBitmap(area)
+                        area.ar_imagem = image_path
                     }
-
-                    saveBitmap(area)
-                    area.ar_imagem = image_path
+                    else{
+                        area.ar_imagem = null;
+                    }
 
                     if (C_Area().updateArea(DBHelper(context), area)){
                         val builder = AlertDialog.Builder(this!!.context!!)
@@ -353,8 +357,14 @@ class AreaCriaAlterDialog : Fragment() {
                     area.setAr_lote_cont(spiContagem_AreaDialog.selectedItemId.toInt())
                     area.ar_nome = edtTxtNome_AreaDialog.text.toString()
                     area.ar_del = "A"
-                    saveBitmap(area)
-                    area.ar_imagem = image_path
+
+                    if(imgArea_DialogFragment.drawable.constantState != resources.getDrawable(R.drawable.ic_menu_camera).constantState){
+                        saveBitmap(area)
+                        area.ar_imagem = image_path
+                    }
+                    else{
+                        area.ar_imagem = null
+                    }
 
                     if (C_Area().insertArea(DBHelper(context), area , userid)){
                         val builder = AlertDialog.Builder(this!!.context!!)
